Improves pluginized build logic
