+++ gradle/plugins/src/main/kotlin/io/dotanuki/norris/gradle/modules/conventions/AndroidPlatformLibraryConventions.kt
@@ -1,0 +1,24 @@
+package io.dotanuki.norris.gradle.modules.conventions
+
+import com.android.build.gradle.BaseExtension
+import io.dotanuki.norris.gradle.modules.models.ProguardRules
+import org.gradle.api.Project
+
+internal fun Project.applyAndroidPlatformLibraryConventions() {
+
+    applyAndroidStandardConventions()
+
+    val android = extensions.findByName("android") as BaseExtension
+
+    android.apply {
+        buildTypes {
+            getByName("release") {
+                isMinifyEnabled = true
+
+                val proguardDefinitions = ProguardRules("$rootDir/app/proguard")
+                proguardFiles(*(proguardDefinitions.extras))
+                proguardFiles(getDefaultProguardFile(proguardDefinitions.androidDefault))
+            }
+        }
+    }
+}
