+++ gradle/plugins/src/main/kotlin/io/dotanuki/norris/gradle/modules/conventions/KotlinConventions.kt
